Add new Submenu component and update parent comps
This component moves the submenu (edit, delete, signout) into it's own comp
